# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI AR Makeup Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/analysis/jobs": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Analysis Job",
        "operationId": "create_analysis_job_api_analysis_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Analysis Job Api Analysis Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/analysis/jobs/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Job",
        "operationId": "get_analysis_job_api_analysis_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Analysis Job Api Analysis Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/reports": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "List Analysis Reports",
        "operationId": "list_analysis_reports_api_analysis_reports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Analysis Reports Api Analysis Reports Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/analysis/reports/{report_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Report",
        "operationId": "get_analysis_report_api_analysis_reports__report_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Analysis Report Api Analysis Reports  Report Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/feedback/jobs": {
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Create Feedback Job",
        "operationId": "create_feedback_job_api_feedback_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Feedback Job Api Feedback Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/feedback/reports": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List Feedback Reports",
        "operationId": "list_feedback_reports_api_feedback_reports_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Feedback Reports Api Feedback Reports Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisJobCreate": {
        "properties": {
          "photoCaptureId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photocaptureid"
          },
          "sourceMediaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcemediaid"
          },
          "previewMediaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previewmediaid"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": "AI makeup analysis"
          },
          "reportTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporttitle"
          },
          "environmentLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environmentlabel"
          },
          "runImmediately": {
            "type": "boolean",
            "title": "Runimmediately",
            "default": false
          },
          "requestPayload": {
            "type": "object",
            "title": "Requestpayload"
          }
        },
        "type": "object",
        "title": "AnalysisJobCreate"
      },
...
      "FeedbackJobCreate": {
        "properties": {
          "photoCaptureId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photocaptureid"
          },
          "uploadedMediaId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadedmediaid"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "camera"
          },
          "sourceLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcelabel"
          },
          "requestPayload": {
            "type": "object",
            "title": "Requestpayload"
          }
        },
        "type": "object",
        "title": "FeedbackJobCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}